Moved to the final structure of the files organization.
Now future_* are now replacing the iterate.py and converge.py.
Updated the .json and the examples
