--- aiida_siesta/workflows/utils/future_converge.py
@@ -1,10 +1,0 @@
-from .converge_absclass import BasicConverger, SequentialConverger
-from .future_iterate import SiestaIterator
-
-
-class SiestaConverger(BasicConverger, SiestaIterator):
-    pass
-
-
-class SiestaSequentialConverger(SequentialConverger):
-    _process_class = SiestaConverger
