+++ setup.json
@@ -48,7 +48,8 @@
 	    "siesta.bandgap = aiida_siesta.workflows.bandgap:BandgapWorkChain",
             "siesta.stm = aiida_siesta.workflows.stm:SiestaSTMWorkChain",
 	    "siesta.iterator = aiida_siesta.workflows.iterate:SiestaIterator",
-	    "siesta.converger = aiida_siesta.workflows.converge:SiestaConverger"
+	    "siesta.converger = aiida_siesta.workflows.converge:SiestaConverger",
+	    "siesta.sequential_converger = aiida_siesta.workflows.converge:SiestaSequentialConverger"
         ],
         "aiida.data": [
             "siesta.psf = aiida_siesta.data.psf:PsfData",
